requirements.txt
# Core PyTorch packages
torch>=2.0.0
torchvision>=0.15.0

# Scientific computing
numpy>=1.24.0
scipy>=1.10.0

# Machine learning utilities
scikit-learn>=1.3.0

# Data visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Data manipulation
pandas>=2.0.0

# Progress bars
tqdm>=4.65.0

# TensorBoard for training visualization
tensorboard>=2.13.0

# Image processing
Pillow>=9.5.0

# For downloading CIFAR-100
urllib3>=2.0.0